Fix security vulnerabilities in webpack
