+++ .eslintrc.js
@@ -1,0 +1,10 @@
+/**
+ * These rules enforce the Hack Reactor Style Guide
+ *
+ * Visit this repo for more information:
+ *   https://github.com/reactorcore/eslint-config-hackreactor
+ */
+
+module.exports = {
+    extends: './node_modules/eslint-config-hackreactor/index.js'
+};
