--- client/index.html
@@ -1,15 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-  <meta charset="UTF-8" />
-  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-  <meta http-equiv="X-UA-Compatible" content="ie=edge" />
-  <title>Regretsy</title>
-</head>
-
-<body>
-  <div id="app"></div>
-</body>
-
-</html>
