+++ client/src/index.js
@@ -1,5 +1,5 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import App from "./app.js";
+import App from "./components/App.js";
 
 ReactDOM.render(<App />, document.getElementById("app"));
