--- client/utilities/auth.js
@@ -1,1 +1,0 @@
-//Filler file
