+++ package.json
@@ -1,73 +1,31 @@
 {
-  "name": "webpack",
+  "name": "regretsy-item-details",
   "version": "1.0.0",
-  "description": "A dope webpack boilerplate",
-  "main": "index.js",
+  "description": "Item details and add to cart component for an Etsy clone",
+  "author": "Brooke Snellings",
+  "license": "",
+  "engines": {
+    "node": ">=6.13.0"
+  },
   "scripts": {
-    "watch": "rimraf dist && webpack --watch",
-    "start": "webpack-dev-server  --open --color ",
-    "build": "rimraf dist &&  webpack --progress --profile",
-    "server": "nodemon server/server.js",
-    "server-prod": "node server/server.js"
+    "react-dev": "webpack -d --watch",
+    "server-dev": "nodemon server"
   },
-  "repository": {
-    "type": "git",
-    "url": "none yet"
-  },
-  "author": "James Dempsey",
-  "license": "UNLICENSED",
   "dependencies": {
     "axios": "^0.18.0",
-    "body-parser": "^1.18.3",
-    "express": "^4.16.4",
+    "express": "^4.17.1",
+    "nodemon": "^1.19.1",
     "mysql": "^2.17.1",
-    "react": "^16.3.2",
-    "react-dom": "^16.3.2",
-    "react-loadable": "^5.4.0",
-    "react-router-dom": "^4.2.2"
-  },
-  "babel": {
-    "presets": [
-      [
-        "env",
-        {
-          "targets": {
-            "browsers": [
-              ">1%",
-              "last 3 versions"
-            ]
-          }
-        }
-      ],
-      "stage-2",
-      "latest",
-      "react"
-    ],
-    "plugins": [
-      "syntax-dynamic-import",
-      "transform-class-properties"
-    ]
+    "react": "^16.8.6",
+    "react-dom": "^16.8.6"
   },
   "devDependencies": {
-    "babel-cli": "^6.26.0",
-    "babel-core": "^6.26.3",
-    "babel-loader": "^7.1.4",
-    "babel-plugin-syntax-dynamic-import": "^6.18.0",
-    "babel-plugin-transform-class-properties": "^6.24.1",
-    "babel-preset-env": "^1.6.1",
-    "babel-preset-latest": "^6.24.1",
-    "babel-preset-react": "^6.24.1",
-    "babel-preset-stage-2": "^6.24.1",
-    "css-loader": "^3.0.0",
-    "file-loader": "^1.1.11",
-    "html-loader": "^0.5.5",
-    "html-webpack-plugin": "^3.2.0",
-    "mini-css-extract-plugin": "^0.4.0",
-    "rimraf": "^2.6.2",
-    "style-loader": "^0.21.0",
-    "uglifyjs-webpack-plugin": "^1.2.5",
-    "webpack": "^4.29.3",
-    "webpack-cli": "^3.2.3",
-    "webpack-dev-server": "^3.1.4"
+    "@babel/core": "^7.4.5",
+    "@babel/preset-env": "^7.4.5",
+    "@babel/preset-react": "^7.0.0",
+    "babel-loader": "^8.0.6",
+    "eslint-config-hackreactor": "git://github.com/reactorcore/eslint-config-hackreactor",
+    "webpack": "^4.35.3",
+    "webpack-cli": "^3.3.5"
   }
 }
